Move a part of AccessibleObject classes to separate files There are no functional changes
