+++ src/System.Windows.Forms/src/System/Windows/Forms/DateTimePicker.cs
@@ -24,7 +24,7 @@ namespace System.Windows.Forms
     [DefaultBindingProperty(nameof(Value))]
     [Designer("System.Windows.Forms.Design.DateTimePickerDesigner, " + AssemblyRef.SystemDesign)]
     [SRDescription(nameof(SR.DescriptionDateTimePicker))]
-    public class DateTimePicker : Control
+    public partial class DateTimePicker : Control
     {
         /// <summary>
         ///  Specifies the default title back color. This field is read-only.
@@ -1672,147 +1672,6 @@ namespace System.Windows.Forms
             }
 
             return new DateTime(s.wYear, s.wMonth, s.wDay, s.wHour, s.wMinute, s.wSecond);
-        }
-
-        private sealed class EnumChildren
-        {
-            public IntPtr hwndFound = IntPtr.Zero;
-
-            public BOOL enumChildren(IntPtr hwnd)
-            {
-                hwndFound = hwnd;
-                return BOOL.TRUE;
-            }
-        }
-
-        public class DateTimePickerAccessibleObject : ControlAccessibleObject
-        {
-            public DateTimePickerAccessibleObject(DateTimePicker owner) : base(owner)
-            {
-            }
-
-            public override string KeyboardShortcut
-            {
-                get
-                {
-                    // APP COMPAT. When computing DateTimePickerAccessibleObject::get_KeyboardShorcut the previous label
-                    // takes precedence over DTP::Text.
-                    // This code was copied from the Everett sources.
-                    Label previousLabel = PreviousLabel;
-
-                    if (previousLabel != null)
-                    {
-                        char previousLabelMnemonic = WindowsFormsUtils.GetMnemonic(previousLabel.Text, false /*convertToUpperCase*/);
-                        if (previousLabelMnemonic != (char)0)
-                        {
-                            return "Alt+" + previousLabelMnemonic;
-                        }
-                    }
-
-                    string baseShortcut = base.KeyboardShortcut;
-
-                    if ((baseShortcut is null || baseShortcut.Length == 0))
-                    {
-                        char ownerTextMnemonic = WindowsFormsUtils.GetMnemonic(Owner.Text, false /*convertToUpperCase*/);
-                        if (ownerTextMnemonic != (char)0)
-                        {
-                            return "Alt+" + ownerTextMnemonic;
-                        }
-                    }
-
-                    return baseShortcut;
-                }
-            }
-
-            public override string Value
-            {
-                get
-                {
-                    string baseValue = base.Value;
-                    if (baseValue is null || baseValue.Length == 0)
-                    {
-                        return Owner.Text;
-                    }
-                    return baseValue;
-                }
-            }
-
-            public override AccessibleStates State
-            {
-                get
-                {
-                    AccessibleStates state = base.State;
-
-                    if (((DateTimePicker)Owner).ShowCheckBox &&
-                       ((DateTimePicker)Owner).Checked)
-                    {
-                        state |= AccessibleStates.Checked;
-                    }
-
-                    return state;
-                }
-            }
-
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    AccessibleRole role = Owner.AccessibleRole;
-                    if (role != AccessibleRole.Default)
-                    {
-                        return role;
-                    }
-
-                    return AccessibleRole.ComboBox;
-                }
-            }
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.IsTogglePatternAvailablePropertyId:
-                        return IsPatternSupported(UiaCore.UIA.TogglePatternId);
-                    case UiaCore.UIA.LocalizedControlTypePropertyId:
-                        return s_dateTimePickerLocalizedControlTypeString;
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.TogglePatternId && ((DateTimePicker)Owner).ShowCheckBox)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            #region Toggle Pattern
-
-            internal override UiaCore.ToggleState ToggleState
-            {
-                get
-                {
-                    return ((DateTimePicker)Owner).Checked ?
-                        UiaCore.ToggleState.On :
-                        UiaCore.ToggleState.Off;
-                }
-            }
-
-            internal override void Toggle()
-            {
-                if (Owner.IsHandleCreated)
-                {
-                    ((DateTimePicker)Owner).Checked = !((DateTimePicker)Owner).Checked;
-                }
-            }
-
-            #endregion
         }
     }
 }
