+++ src/System.Windows.Forms/src/System/Windows/Forms/NumericUpDown.cs
@@ -19,7 +19,7 @@ namespace System.Windows.Forms
     [DefaultEvent(nameof(ValueChanged))]
     [DefaultBindingProperty(nameof(Value))]
     [SRDescription(nameof(SR.DescriptionNumericUpDown))]
-    public class NumericUpDown : UpDownBase, ISupportInitialize
+    public partial class NumericUpDown : UpDownBase, ISupportInitialize
     {
         private const decimal DefaultValue = decimal.Zero;
         private const decimal DefaultMinimum = decimal.Zero;
@@ -884,99 +884,6 @@ namespace System.Windows.Forms
             }
             Debug.Assert(largestDigit != -1 && digitWidth != -1, "Failed to find largest digit.");
             return largestDigit;
-        }
-
-        internal class NumericUpDownAccessibleObject : ControlAccessibleObject
-        {
-            private readonly UpDownBase _owner;
-
-            public NumericUpDownAccessibleObject(NumericUpDown owner) : base(owner)
-            {
-                _owner = owner;
-            }
-
-            public override AccessibleObject GetChild(int index)
-            {
-                // TextBox child
-                if (index == 0)
-                {
-                    return _owner.TextBox.AccessibilityObject.Parent;
-                }
-
-                // Up/down buttons
-                if (index == 1)
-                {
-                    return _owner.UpDownButtonsInternal.AccessibilityObject.Parent;
-                }
-
-                return null;
-            }
-
-            public override int GetChildCount()
-            {
-                return 2;
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.RuntimeIdPropertyId:
-                        return RuntimeId;
-                    case UiaCore.UIA.NamePropertyId:
-                        return Name;
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.SpinnerControlTypeId;
-                    case UiaCore.UIA.BoundingRectanglePropertyId:
-                        return Bounds;
-                    case UiaCore.UIA.LegacyIAccessibleStatePropertyId:
-                        return State;
-                    case UiaCore.UIA.LegacyIAccessibleRolePropertyId:
-                        return Role;
-                    case UiaCore.UIA.IsKeyboardFocusablePropertyId:
-                        return false;
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    AccessibleRole role = Owner.AccessibleRole;
-
-                    if (role != AccessibleRole.Default)
-                    {
-                        return role;
-                    }
-
-                    return AccessibleRole.SpinButton;
-                }
-            }
-
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    if (_owner is null)
-                    {
-                        return base.RuntimeId;
-                    }
-
-                    // we need to provide a unique ID
-                    // others are implementing this in the same manner
-                    // first item is static - 0x2a (RuntimeIDFirstItem)
-                    // second item can be anything, but here it is a hash
-
-                    var runtimeId = new int[3];
-                    runtimeId[0] = RuntimeIDFirstItem;
-                    runtimeId[1] = (int)(long)_owner.InternalHandle;
-                    runtimeId[2] = _owner.GetHashCode();
-
-                    return runtimeId;
-                }
-            }
         }
     }
 }
