+++ src/System.Windows.Forms/src/System/Windows/Forms/ProgressBar.cs
@@ -19,7 +19,7 @@ namespace System.Windows.Forms
     [DefaultProperty(nameof(Value))]
     [DefaultBindingProperty(nameof(Value))]
     [SRDescription(nameof(SR.DescriptionProgressBar))]
-    public class ProgressBar : Control
+    public partial class ProgressBar : Control
     {
         // These four values define the range of possible values, how to navigate through them and the
         // current position
@@ -706,68 +706,5 @@ namespace System.Windows.Forms
         /// </returns>
         protected override AccessibleObject CreateAccessibilityInstance()
             => new ProgressBarAccessibleObject(this);
-
-        internal class ProgressBarAccessibleObject : ControlAccessibleObject
-        {
-            internal ProgressBarAccessibleObject(ProgressBar owner) : base(owner)
-            {
-            }
-
-            private ProgressBar OwningProgressBar => Owner as ProgressBar;
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.ValuePatternId ||
-                    patternId == UiaCore.UIA.RangeValuePatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.NamePropertyId:
-                        return Name;
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.ProgressBarControlTypeId;
-                    case UiaCore.UIA.IsKeyboardFocusablePropertyId:
-                        // This is necessary for compatibility with MSAA proxy:
-                        // IsKeyboardFocusable = true regardless the control is enabled/disabled.
-                        return true;
-                    case UiaCore.UIA.IsRangeValuePatternAvailablePropertyId:
-                    case UiaCore.UIA.IsValuePatternAvailablePropertyId:
-                    case UiaCore.UIA.RangeValueIsReadOnlyPropertyId:
-                        return true;
-                    case UiaCore.UIA.RangeValueLargeChangePropertyId:
-                    case UiaCore.UIA.RangeValueSmallChangePropertyId:
-                        return double.NaN;
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
-
-            internal override void SetValue(double newValue)
-            {
-                throw new InvalidOperationException("Progress Bar is read-only.");
-            }
-
-            internal override double LargeChange => double.NaN;
-
-            internal override double Maximum => OwningProgressBar?.Maximum ?? double.NaN;
-
-            internal override double Minimum => OwningProgressBar?.Minimum ?? double.NaN;
-
-            internal override double SmallChange => double.NaN;
-
-            internal override double RangeValue => OwningProgressBar?.Value ?? double.NaN;
-
-            internal override bool IsReadOnly => true;
-        }
     }
 }
