+++ src/System.Windows.Forms/src/System/Windows/Forms/StatusStrip.cs
@@ -14,7 +14,7 @@ using static Interop;
 namespace System.Windows.Forms
 {
     [SRDescription(nameof(SR.DescriptionStatusStrip))]
-    public class StatusStrip : ToolStrip
+    public partial class StatusStrip : ToolStrip
     {
         private const AnchorStyles AllAnchor = AnchorStyles.Left | AnchorStyles.Right | AnchorStyles.Bottom | AnchorStyles.Top;
         private const AnchorStyles HorizontalAnchor = AnchorStyles.Left | AnchorStyles.Right;
@@ -644,78 +644,6 @@ namespace System.Windows.Forms
                 }
                 base.WndProc(ref m);
             }
-        }
-
-        internal class StatusStripAccessibleObject : ToolStripAccessibleObject
-        {
-            public StatusStripAccessibleObject(StatusStrip owner) : base(owner)
-            {
-            }
-
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    AccessibleRole role = Owner.AccessibleRole;
-                    if (role != AccessibleRole.Default)
-                    {
-                        return role;
-                    }
-                    return AccessibleRole.StatusBar;
-                }
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                if (propertyID == UiaCore.UIA.ControlTypePropertyId)
-                {
-                    return UiaCore.UIA.StatusBarControlTypeId;
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
-
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                if (!(Owner is StatusStrip statusStrip) || statusStrip.Items.Count == 0)
-                {
-                    return null;
-                }
-
-                switch (direction)
-                {
-                    case UiaCore.NavigateDirection.FirstChild:
-                        AccessibleObject firstChild = null;
-                        for (int i = 0; i < GetChildCount(); i++)
-                        {
-                            firstChild = GetChild(i);
-                            if (firstChild != null && !(firstChild is ControlAccessibleObject))
-                            {
-                                return firstChild;
-                            }
-                        }
-                        return null;
-
-                    case UiaCore.NavigateDirection.LastChild:
-                        AccessibleObject lastChild = null;
-                        for (int i = GetChildCount() - 1; i >= 0; i--)
-                        {
-                            lastChild = GetChild(i);
-                            if (lastChild != null && !(lastChild is ControlAccessibleObject))
-                            {
-                                return lastChild;
-                            }
-                        }
-                        return null;
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            internal override UiaCore.IRawElementProviderFragment ElementProviderFromPoint(double x, double y)
-                => Owner.IsHandleCreated ? HitTest((int)x, (int)y) : null;
-
-            internal override UiaCore.IRawElementProviderFragment GetFocus() => Owner.IsHandleCreated ? GetFocused() : null;
         }
     }
 }
