+++ src/System.Windows.Forms/src/System/Windows/Forms/ToolStripOverflowButton.ToolStripOverflowButtonAccessibleObject.cs
@@ -1,0 +1,50 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable disable
+
+using static Interop;
+
+namespace System.Windows.Forms
+{
+    public partial class ToolStripOverflowButton
+    {
+        internal class ToolStripOverflowButtonAccessibleObject : ToolStripDropDownItemAccessibleObject
+        {
+            private string stockName;
+
+            public ToolStripOverflowButtonAccessibleObject(ToolStripOverflowButton owner) : base(owner)
+            {
+            }
+
+            public override string Name
+            {
+                get
+                {
+                    string name = Owner.AccessibleName;
+                    if (name != null)
+                    {
+                        return name;
+                    }
+                    if (string.IsNullOrEmpty(stockName))
+                    {
+                        stockName = SR.ToolStripOptions;
+                    }
+                    return stockName;
+                }
+                set => base.Name = value;
+            }
+
+            internal override object GetPropertyValue(UiaCore.UIA propertyID)
+            {
+                if (propertyID == UiaCore.UIA.ControlTypePropertyId)
+                {
+                    return UiaCore.UIA.MenuItemControlTypeId;
+                }
+
+                return base.GetPropertyValue(propertyID);
+            }
+        }
+    }
+}
