+++ src/System.Windows.Forms/src/System/Windows/Forms/ToolStripOverflowButton.cs
@@ -7,7 +7,6 @@
 using System.ComponentModel;
 using System.Drawing;
 using System.Windows.Forms.Design;
-using static Interop;
 
 namespace System.Windows.Forms
 {
@@ -15,7 +14,7 @@ namespace System.Windows.Forms
     ///  ToolStripOverflowButton
     /// </summary>
     [ToolStripItemDesignerAvailability(ToolStripItemDesignerAvailability.None)]
-    public class ToolStripOverflowButton : ToolStripDropDownButton
+    public partial class ToolStripOverflowButton : ToolStripDropDownButton
     {
         // we need to cache this away as the Parent property gets reset a lot.
         private readonly ToolStrip parentToolStrip;
@@ -141,43 +140,6 @@ namespace System.Windows.Forms
             {
                 ToolStripRenderer renderer = ParentInternal.Renderer;
                 renderer.DrawOverflowButtonBackground(new ToolStripItemRenderEventArgs(e.Graphics, this));
-            }
-        }
-
-        internal class ToolStripOverflowButtonAccessibleObject : ToolStripDropDownItemAccessibleObject
-        {
-            private string stockName;
-
-            public ToolStripOverflowButtonAccessibleObject(ToolStripOverflowButton owner) : base(owner)
-            {
-            }
-
-            public override string Name
-            {
-                get
-                {
-                    string name = Owner.AccessibleName;
-                    if (name != null)
-                    {
-                        return name;
-                    }
-                    if (string.IsNullOrEmpty(stockName))
-                    {
-                        stockName = SR.ToolStripOptions;
-                    }
-                    return stockName;
-                }
-                set => base.Name = value;
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                if (propertyID == UiaCore.UIA.ControlTypePropertyId)
-                {
-                    return UiaCore.UIA.MenuItemControlTypeId;
-                }
-
-                return base.GetPropertyValue(propertyID);
             }
         }
     }
